requirements.txt
Django==4.2.1
djangorestframework==3.14.0
pandas==1.5.3
keras==2.10.0
tensorflow==2.10.0
numpy==1.24.3
Pillow==9.5.0
absl-py==1.3.0